server: Use memfd to back anonymous mappings on Linux.
On Linux, try creating a memfd first to implement create_temp_file. If a large page mapping is
requested (SEC_LARGE_PAGES), create a memfd with the MFD_HUGETLB flags. If SEC_COMMIT is passed,
allocate sufficient space to back the mapping with posix_fallocate.

Additionally, ensure when SEC_LARGE_PAGES is requested, ensure that SEC_COMMIT is passed as well.