winemac.drv: Use ScreenCaptureKit for snapshotting windows when possible.
The CG methods for capturing screen contents are deprecated and
trigger a permissions prompt in Sequoia. ScreenCaptureKit's
shareable content APIs are the replacement, and particularly
-getCurrentProcessShareableContentWithCompletionHandler: does not
require any permissions for capturing images within the same app.

See 496b001ae0b for why we have to do this manually in the first
place.