ntdll: Exclude biggest envars to fit 32767 chars limit in get_initial_environment.
Windows has a limit of 32767 characters in the environment block. https://superuser.com/questions/1070272/why-does-windows-have-a-limit-on-environment-variables-at-all

Wine doesn't attempt to respect that at all. Do some programs actually depend on this? Unfortunately yes.

With this patch, during initialization, if the block size would exceed (or be close to) the limit, the biggest environment variables will be excluded.

This is useful when a user has very long environment variables in their system for reasons unrelated to Wine.

Do note that there's still nothing done to make sure that the limit isn't exceeded after initialization.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=56941
Signed-off-by: Martino Fontana <tinozzo123@gmail.com>