winex11.drv: Properly check if the desktop window is fullscreen on multi-monitor systems.
If there are two monitors and the primary monitor is on the right side and the virtual desktop
resolution happens to be the same as the primary monitor, then the X11 desktop window will
not have window frames and show as a fullscreen window before this patch. is_desktop_fullscreen()
should check which monitor that it is on to decide whether to use fullscreen mode instead of always
checking against the primary monitor.