explorer: Support the NoTrayItemsDisplay registry setting.
This setting is a more extreme version of the ShowSystray setting: In
addition to suppressing the floating systray, it suppresses systray
icons in the host taskbar and in the virtual desktop taskbar.
NoTrayItemsDisplay is the same registry setting that can be set on
Windows to remove the system tray from the taskbar, so if any
application sets it, Wine will now respect the setting.