comdlg32: Allow entering a filter in the itemdlg file name field.
Current filter is initially empty and can be set by typing a filter
into the file name field or selecting an entry from the file type menu.

Selecting a file type entry with an empty spec leaves the current filter unchanged.

This behavior matches windows more closely and fixes applications
that rely on SetFileName() for filtering.

Signed-off-by: Charlotte Pabst <cpabst@codeweavers.com>