win32u: Make sure to load drivers when updating the display cache.
If the registry is initialized, a process calling ChangeDisplaySettings
will only load the driver when calling GetCurrentDisplaySettings in
source_get_full_mode / source_get_current_settings.

As we're going to cache the current display mode, this will not be
called anymore and the driver won't be loaded at all. So, make sure we
load it earlier instead.