wintab32: Align WTPACKET for 32/64-bit archs.
WTPACKET's structure is never directly exposed via the API; it's
internal to Wine. The HCTX value is only used on the wintab32 side,
not the driver side, and it is used as an integer rather than a
pointer, so this should be safe to do.

This eliminates the need to have a wow64 thunk for tablet_get_packet.