mshtml: Fix builtin style translation in removeAttribute for IE9+ modes.
Because we're supposed to translate styles here to what Gecko expects via the
style table, we don't actually need to look up the *real* prototypes on the
object (what it has currently in js code), but rather just the builtin list
of style props of the object, which must include all the styles supported. So
we look into the entire builtin prototype chain.

remove_attribute also only deals with builtin dispids, so it makes no sense
to use it on the output of GetDispID in IE9+ modes. Legacy modes do remove
all props though, added tests to confirm this.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>