winex11: Use the style to decide whether a window is currently visible or iconic.
The mapped and iconic flags should match the last requested state at
all time. Unless we've missed sending an update this should match the
current Win32 state, and we should be able use it instead as a more
accurate source of information about the supposed window state.

When receiving config changes, we don't want to apply changes to a
window which is supposed to be hidden or minimized. When receiving state
changes, we don't want to apply changes to a window which is supposed to
be hidden, regardless of the X11 believes.