wined3d: Do not clamp fog in the VS.
NVidia and WARP clamp before interpolation in the oFog case; AMD clamps only
after interpolation (which we also do).

On the other hand, test_fog() proves that clamping should not be done before
interpolation in the FFP case, for any driver.

In order to emulate an FFP vertex shader with a real compiled shader, we need to
avoid clamping before interpolation, choosing the AMD behaviour here.